Script for Generative intialization
For stability reasons it's important to initialize the network
such that it's monotonic and has support for the whole output space.

Due to nonlinear activation functions this is probably impossible to do
by just initializing the weights so we train the model to approximate
the identity function (modulo the prior CDF).